Moved and renamed components to align with design guidelines
