+++ packages/venia-ui/lib/components/SearchPage/categoryFilters.css
