+++ packages/venia-ui/lib/components/SearchPage/index.js
@@ -1,0 +1,1 @@
+export { default } from './searchPage';
