+++ packages/venia-ui/lib/components/SearchPage/searchPage.css
