+++ packages/venia-ui/lib/components/SearchPage/searchPage.js
@@ -1,0 +1,119 @@
+import React, { useCallback, useEffect } from 'react';
+import { object, shape, string } from 'prop-types';
+import { useQuery } from '@apollo/react-hooks';
+import { useAppContext } from '@magento/peregrine/lib/context/app';
+
+import { mergeClasses } from '../../classify';
+import Gallery from '../Gallery';
+import FilterModal from '../FilterModal';
+import { fullPageLoadingIndicator } from '../LoadingIndicator';
+import PRODUCT_SEARCH from '../../queries/productSearch.graphql';
+import getQueryParameterValue from '../../util/getQueryParameterValue';
+import CategoryFilters from './categoryFilters';
+import defaultClasses from './searchPage.css';
+
+const SearchPage = props => {
+    const { history, location } = props;
+    const classes = mergeClasses(defaultClasses, props.classes);
+
+    // retrieve app state and action creators
+    const [appState, appApi] = useAppContext();
+    const { searchOpen } = appState;
+    const { executeSearch, toggleDrawer, toggleSearch } = appApi;
+
+    const openDrawer = useCallback(() => {
+        toggleDrawer('filter');
+    }, [toggleDrawer]);
+
+    // get the URL query parameters.
+    const urlQueryValue = getQueryParameterValue({
+        location,
+        queryParameter: 'query'
+    });
+    const categoryId = getQueryParameterValue({
+        location,
+        queryParameter: 'category'
+    });
+
+    // derive initial state from query params
+    // never re-run this effect, even if deps change
+    /* eslint-disable react-hooks/exhaustive-deps */
+    useEffect(() => {
+        // ensure search is open to begin with
+        if (toggleSearch && !searchOpen && urlQueryValue) {
+            toggleSearch();
+        }
+    }, []);
+    /* eslint-enable react-hooks/exhaustive-deps */
+
+    const apolloQueryVariable = categoryId
+        ? { inputText: urlQueryValue, categoryId }
+        : { inputText: urlQueryValue };
+
+    const { loading, error, data } = useQuery(PRODUCT_SEARCH, {
+        variables: apolloQueryVariable
+    });
+
+    if (loading) return fullPageLoadingIndicator;
+    if (error) {
+        return (
+            <div className={classes.noResult}>
+                No results found. The search term may be missing or invalid.
+            </div>
+        );
+    }
+
+    const { products } = data;
+    const { filters, total_count, items } = products;
+
+    if (items.length === 0) {
+        return <div className={classes.noResult}>No results found!</div>;
+    }
+
+    const maybeCategoryFilters = categoryId ? (
+        <CategoryFilters
+            categoryId={categoryId}
+            executeSearch={executeSearch}
+            history={history}
+            location={location}
+        />
+    ) : null;
+
+    const maybeFilterButtons = filters ? (
+        <div className={classes.headerButtons}>
+            <button onClick={openDrawer} className={classes.filterButton}>
+                Filter
+            </button>
+        </div>
+    ) : null;
+
+    const maybeFilterModal = filters ? <FilterModal filters={filters} /> : null;
+
+    return (
+        <article className={classes.root}>
+            <div className={classes.categoryTop}>
+                <div className={classes.totalPages}>
+                    {`${total_count} items`}
+                </div>
+                {maybeCategoryFilters}
+                {maybeFilterButtons}
+            </div>
+            {maybeFilterModal}
+            <section className={classes.gallery}>
+                <Gallery items={items} />
+            </section>
+        </article>
+    );
+};
+
+export default SearchPage;
+
+SearchPage.propTypes = {
+    classes: shape({
+        noResult: string,
+        root: string,
+        totalPages: string
+    }),
+    history: object,
+    location: object.isRequired
+};
